Fix corrupt container, add deploy sanity checks

+++ scripts/update-version.js
@@ -4,6 +4,7 @@
 const args = require('minimist')(process.argv.slice(2), { boolean: true })
 const chalk = require('chalk')
 const fs   = require('fs')
+const { runMain } = require('module')
 
 const filename = __dirname + '/../package.json'
 if (!args.t) {
@@ -13,34 +14,43 @@ if (!args.t) {
 
 console.log(args);
 
-try {
-    const package = require(filename)
-    fs.writeFileSync(filename, JSON.stringify({...package, version: args.t}, null, 2) , 'utf-8')
-    console.log(chalk.green(`✓ Successfully updated package.json version to ${args.t}.`))
-} catch (err) {
-    console.error(chalk.red(`✗ Could not updated package.json version to ${args.t}.`))
-    console.log(err) 
-    process.exit(1)
-}
+(async _=> {
+	try {
+		const package = require(filename)
+		fs.writeFileSync(filename, JSON.stringify({...package, version: args.t}, null, 2) , 'utf-8')
+		console.log(chalk.green(`✓ Successfully updated package.json version to ${args.t}.`))
+	} catch (err) {
+		console.error(chalk.red(`✗ Could not updated package.json version to ${args.t}.`))
+		console.log(err) 
+		process.exit(1)
+	}
 
-let sqlUpdateFile = `Changes/${args.t}/db-updates-${args.t}.sql`
-try {
-    let sqlString = fs.readFileSync(sqlUpdateFile, 'utf8')
-    if (!sqlString.startsWith('START TRANSACTION;') && args['wrap-SQL']) {
-        // Wrap the SQL string with code required to update a version
-        sqlString = 'START TRANSACTION;\n' + sqlString + 
-                    `SELECT @VER := "${args.t}";\n` +
-                    'INSERT INTO `db_version` (version, completed) VALUES (@VER, current_timestamp());\n' +
-                    'COMMIT;'
-        console.log(chalk.green(`✓ Wrapped SQL at ${sqlUpdateFile}`));
-    } else {
-        sqlString = sqlString.replace(/\$CURRENT_DATABASE_VERSION/g, args.t);
-    }
-    fs.writeFileSync(sqlUpdateFile, sqlString, 'utf-8')
-    console.log(chalk.green(`✓ Successfully updated ${sqlUpdateFile} database version to ${args.t}.`))    
-} catch (err) {
-    console.error(chalk.red(`✗ Could not updated the file ${sqlUpdateFile} to database version to ${args.t}.`))
-    console.log(err)
-}
+	let sqlUpdateFile = `Changes/${args.t}/db-updates-${args.t}.sql`
 
-process.exit(0)
+	try {
+		if (!fs.existsSync(sqlUpdateFile)) throw 'File not found:' + sqlUpdateFile;				// if the sql script file is gzipped, this would've exploded
+
+		let sqlString = fs.readFileSync(sqlUpdateFile, 'utf8')
+		if (!sqlString.startsWith('START TRANSACTION;') && args['wrap-SQL']) {
+			// Wrap the SQL string with code required to update a version
+			sqlString = 'START TRANSACTION;\n' + sqlString + 
+						`SELECT @VER := "${args.t}";\n` +
+						'INSERT INTO `db_version` (version, completed) VALUES (@VER, current_timestamp());\n' +
+						'COMMIT;'
+			console.log(chalk.green(`✓ Wrapped SQL at ${sqlUpdateFile}`));
+		} else {
+			sqlString = sqlString.replace(/\$CURRENT_DATABASE_VERSION/g, args.t);
+		}
+		fs.writeFileSync(sqlUpdateFile, sqlString, 'utf-8')
+		
+		console.log('Zipping the SQL file...');
+		await run('gzip', sqlUpdateFile);
+		
+		console.log(chalk.green(`✓ Successfully updated ${sqlUpdateFile} database version to ${args.t}.`))    
+	} catch (err) {
+		console.error(chalk.red(`✗ Could not update the file ${sqlUpdateFile} to database version to ${args.t}.`))
+		console.log(err)
+	}
+
+	process.exit(0)
+})();
